requirements.txt
transformers==4.46.3
fastapi==0.109.0
uvicorn==0.24.0
pymongo==4.6.1
python-multipart==0.0.9
torch==2.2.2
torchvision==0.17.2
pillow==10.2.0
assemblyai==0.17.0
numpy<2.0
pydantic-settings>=0.1.0
requests==2.31.0
pydantic==2.6.4